requirements.txt
customtkinter>=5.2.0
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
py2app>=0.28.0
